+++ src/main/java/repository/ShowRepository.java
@@ -9,7 +9,7 @@ public class ShowRepository extends Repository<Show> {
         super(entityManager);
     }
 
-    public Show getById(int Id) {
+    public Show getById(long Id) {
         return em.find(Show.class, Id);
     }
 }
